fix: 代码审查发现的严重问题修复
修复问题清单：

【严重Bug】
1. ChunkedDownloader: 修复进度回退逻辑错误 - bytesCounter 未在回调中更新，
   导致下载失败回退时无法正确减少已下载字节数
2. SpeedStat: 修复 sampleSpeed() 除零风险 - 当时间跨度为0时直接返回0
3. SpeedStat: 修复 feed() 清理逻辑错误 - 旧实现可能误删有效数据，
   改用从尾部逐个清理过期数据的方式
4. AlistProtocol: 修复 requestText() 和 downloadFile() 空实现问题 -
   之前 requestText() 直接返回空字符串，downloadFile() 不做任何操作
5. AlistProtocol: 修复 fetchDownloadLink() JSON body 格式错误 -
   缺少 {} 包裹导致无效 JSON

【中等问题】
6. HttpProtocol: 修复 requestText() Response 资源泄漏 -
   未使用 try-with-resources 关闭 Response
7. ChunkedDownloader: 修复 createChunks() 整数除法精度丢失 -
   改用均匀分配算法避免最后一个分片过大
8. ParallelDownloader: 空文件下载时增加修改时间设置
9. McpatchProtocol: 修复 requestText() 整数溢出风险 -
   len > Integer.MAX_VALUE 时直接抛出异常
10. ParallelDownloader: 线程池改为跨重试轮次复用，避免每轮重建
11. Servers: 修复 multipleAvailableServers() 竞态条件 -
    使用 CAS 基于实际读取的 serverIndex 推进索引
12. ReduceReportingFrequency: feed() 参数类型从 int 改为 long

【轻微问题】
13. AlistProtocol: 修复拼写错误 '连接超市' -> '连接超时'
14. Crc64_XZ/Crc16_IBM_SDLC: 移除 reset() 中不必要的 buf 清零
15. 移除 HashUtility 中不再使用的 Arrays import

+++ src/main/java/com/github/balloonupdate/mcpatch/client/utils/ReduceReportingFrequency.java
@@ -14,7 +14,7 @@ public class ReduceReportingFrequency {
      */
     private long accumulated = 0L;
 
-    public long feed(int bytes) {
+    public long feed(long bytes) {
         accumulated += bytes;
 
         long now = System.currentTimeMillis();
